+++ src/app/card/[id]/page.tsx
@@ -1,5 +1,9 @@
 import { notFound } from 'next/navigation';
 import { supabase } from '@/utils/supabase/createClient';
+import { convertToCamelCase } from '@/utils/convert/invitaitonTypeConvert';
+import MainPhoto from '@/components/card/MainPhoto';
+import Greeting from '@/components/card/Greeting';
+import PersonalInfoOnSharedCard from '@/components/card/PersonalInfoOnSharedCard';
 
 export const generateStaticParams = async () => {
   const { data } = await supabase.from('invitation').select('id');
@@ -19,9 +23,42 @@ const fetchInvitationData = async (id: string) => {
 };
 
 const CardPage = async ({ params }: { params: { id: string } }) => {
-  const { isPrivate } = await fetchInvitationData(params.id);
+  const invitation = await fetchInvitationData(params.id);
+  const {
+    // gallery,
+    // type,
+    // mood,
+    mainView,
+    bgColor,
+    stickers,
+    // imgRatio,
+    // mainText,
+    greetingMessage,
+    // guestbook,
+    // attendance,
+    personalInfo,
+    // weddingInfo,
+    // account,
+    // navigationDetail,
+    // dDay,
+    mainPhotoInfo,
+    isPrivate,
+  } = convertToCamelCase(invitation);
 
-  return <div>{isPrivate ? <div>아직 공개되지 않은 청첩장입니다.</div> : <div>청첩장 내용</div>}</div>;
+  return isPrivate ? (
+    <div>아직 공개되지 않은 청첩장입니다.</div>
+  ) : (
+    <div>
+      <MainPhoto
+        mainPhotoInfo={mainPhotoInfo}
+        bgColor={bgColor}
+        mainView={mainView}
+        stickers={stickers}
+      />
+      <Greeting greetingMessage={greetingMessage} />
+      <PersonalInfoOnSharedCard personalInfo={personalInfo} />
+    </div>
+  );
 };
 
 export default CardPage;
