+++ src/components/create/preview/GreetingPreview.tsx
@@ -1,5 +1,6 @@
 import { InvitationFormType } from '@/types/invitationFormType.type';
 import { Control, useWatch } from 'react-hook-form';
+import Greeting from '@/components/card/Greeting';
 
 const GreetingPreview = ({ control }: { control: Control<InvitationFormType> }) => {
   const greetingMessage = useWatch({
@@ -7,16 +8,7 @@ const GreetingPreview = ({ control }: { control: Control<InvitationFormType> }) 
     name: 'greetingMessage',
   });
 
-  return (
-    <div className='text-black  flex flex-col justify-center items-center'>
-      {!greetingMessage.title ? <p>제목을 입력해주세요</p> : <p>{greetingMessage.title}</p>}
-      <div
-        dangerouslySetInnerHTML={{
-          __html: greetingMessage?.content || '대표문구를 작성해주세요',
-        }}
-      ></div>
-    </div>
-  );
+  return <Greeting greetingMessage={greetingMessage} />;
 };
 
 export default GreetingPreview;
